refactor: hooks/useCollisionDetection
