+++ client/features/playground/preview/Preview.tsx
@@ -7,6 +7,7 @@ import { Controls } from './components/Controls/Controls';
 import { Goal } from './components/Goal/Goal';
 import { Obstacle } from './components/Obstacle/Obstacle';
 import { Sprite } from './components/Sprite/Sprite';
+import { useCollisionDetection } from './hooks/useCollisionDetection';
 
 type Props = {
   scripts: Scripts;
@@ -23,36 +24,31 @@ const defaultScriptState = (script: Block[]) => ({
 
 export const Preview = (props: Props) => {
   const { scripts } = props;
+
+  // ステート管理
   const [stepSpeed, setStepSpeed] = useState(1);
-  const [hasReachedGoal, setHasReachedGoal] = useState(false);
-  const [collisions, setCollisions] = useState<boolean>(false);
-  const [obstacles] = useState([
-    { x: 100, y: 100, width: 50, height: 50 },
-    { x: 200, y: 200, width: 50, height: 50 },
-  ]);
 
   const [scriptStates, setScriptStates] = useState<ScriptState[]>(
     scripts?.map(({ script }) => defaultScriptState(script)),
   );
+
   const [state, setState] = useState<SpriteState>({
     x: 0,
     y: 0,
     direction: 0,
   });
 
-  const goal = {
-    x: 590,
-    y: 480,
-    width: 50,
-    height: 50,
-  };
+  // カスタムフックによる衝突とゴール判定
+  const { hasReachedGoal, collisions } = useCollisionDetection(state);
 
+  // スクリプト状態更新関数
   const updateScriptState = (updateFn: (newScriptStates: ScriptState[]) => void) => {
     const newScriptStates = structuredClone(scriptStates);
     updateFn(newScriptStates);
     setScriptStates([...newScriptStates]);
   };
 
+  // スクリプト実行関数
   const interval = (i: number, script: Block[], stepCount: number[]) => {
     if (stepCount[0] >= script.length) {
       updateScriptState((scriptStates) => {
@@ -64,6 +60,7 @@ export const Preview = (props: Props) => {
     const block = script?.[stepCount[0]];
 
     if (block === undefined) return;
+
     updateScriptState((scriptStates) => {
       const step = (block: blockArg): void | string | undefined => {
         if (typeof block === 'string') {
@@ -77,6 +74,7 @@ export const Preview = (props: Props) => {
     });
   };
 
+  // インターバルID管理関数
   const intervalId = ({ script, active, stepDelay, stepCount }: ScriptState, i: number) => {
     if (active) {
       return setInterval(() => interval(i, script, stepCount), (stepDelay ?? stepSpeed) * 1000);
@@ -84,66 +82,25 @@ export const Preview = (props: Props) => {
     return undefined;
   };
 
+  // スクリプト実行タイミングの管理（useEffectフック）
   useEffect(() => {
     const intervalIds = scriptStates.map(intervalId);
-    return () => {
-      intervalIds?.forEach((intervalId) => clearInterval(intervalId));
-    };
+
+    return () => intervalIds?.forEach((intervalId) => clearInterval(intervalId));
   }, [scriptStates, stepSpeed]);
 
+  // スタートボタン押下時の処理
   const handleStartButtonClick = () => {
     setScriptStates(
       scriptStates
         .filter(({ script }) => script[0]?.id === 0)
-        .map((scriptState) => ({ ...scriptState, active: !scriptState.active, stepDelay: 0 })),
+        .map((scriptState) => ({
+          ...scriptState,
+          active: !scriptState.active,
+          stepDelay: 0,
+        })),
     );
   };
-
-  useEffect(() => {
-    const spriteSize = 30;
-    //eslint-disable-next-line complexity
-    const checkCollisionAndGoal = () => {
-      const spriteRect = {
-        left: state.x,
-        right: state.x + spriteSize,
-        top: state.y,
-        bottom: state.y + spriteSize,
-      };
-
-      const hasCollision = obstacles.some(
-        (obstacle) =>
-          !(
-            spriteRect.left > obstacle.x + obstacle.width ||
-            spriteRect.right < obstacle.x ||
-            spriteRect.top > obstacle.y + obstacle.height ||
-            spriteRect.bottom < obstacle.y
-          ),
-      );
-
-      const reachedGoal = !(
-        spriteRect.left > goal.x + goal.width ||
-        spriteRect.right < goal.x ||
-        spriteRect.top > goal.y + goal.height ||
-        spriteRect.bottom < goal.y
-      );
-
-      if (hasCollision) {
-        setCollisions(true);
-        setState({ x: 0, y: 0, direction: 0 });
-        setScriptStates((prevStates) => prevStates.map((state) => ({ ...state, active: false })));
-      } else {
-        setCollisions(false);
-      }
-
-      if (reachedGoal) {
-        setHasReachedGoal(true);
-        setScriptStates((prevStates) => prevStates.map((state) => ({ ...state, active: false })));
-      }
-    };
-
-    const checkInterval = setInterval(checkCollisionAndGoal, 100);
-    return () => clearInterval(checkInterval);
-  }, [state, obstacles]);
 
   return (
     <div className={styles.main}>
