+++ client/features/playground/preview/constants.ts
@@ -1,0 +1,6 @@
+export const goal = {
+  x: 590,
+  y: 500,
+  width: 50,
+  height: 50,
+};
